each post is has the username associated with it
