+++ backend/middlewares/AuthMiddleware.js
@@ -1,0 +1,20 @@
+const jwt = require('jsonwebtoken')
+
+const verifyJWT = (req, res, next) => {
+    const token = req.headers["x-access-token"]
+
+    if(!token) {
+        res.send("Yo, we need a token, please give it next time")
+    } else {
+        jwt.verify(token, "tochange", (err, decoded) => {
+            if(err){
+                res.json({auth: false, message: "failed auth" })
+            }else{
+                req.user = decoded
+                next()
+            }
+        })
+    }
+}
+
+module.exports = {verifyJWT}
