+++ backend/models/Posts.js
@@ -12,7 +12,7 @@ module.exports = (sequelize, DataTypes) => {
         },
         username: {
             type: DataTypes.STRING,
-            allowNull: true,
+            allowNull: false,
         },
         
     })
