+++ backend/models/Users.js
@@ -21,5 +21,13 @@ module.exports = (sequelize, DataTypes) => {
         
     })
 
+    Users.associate = (models) => {
+        Users.hasMany(models.Posts, {
+            onDelete: "cascade",
+        })
+    }
+
+
+
     return Users
 }
